# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 61be5abf7a739847d5cfd2d6
  title: Cashback
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /produtos:
    post:
      x-skaffolder-id: 61be9d727a739847d5cfd39f
      x-skaffolder-name: create
      x-skaffolder-id-resource: 61be9d077a739847d5cfd390
      x-skaffolder-resource: Produto
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Produto
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
    get:
      x-skaffolder-id: 61be9d727a739847d5cfd3a5
      x-skaffolder-name: list
      x-skaffolder-id-resource: 61be9d077a739847d5cfd390
      x-skaffolder-resource: Produto
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Produto
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produto"
  "/produtos/{id}":
    delete:
      x-skaffolder-id: 61be9d727a739847d5cfd3a6
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61be9d077a739847d5cfd390
      x-skaffolder-resource: Produto
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Produto
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61be9d727a739847d5cfd3a2
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61be9d077a739847d5cfd390
      x-skaffolder-resource: Produto
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Produto
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
    post:
      x-skaffolder-id: 61be9d727a739847d5cfd3a0
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61be9d077a739847d5cfd390
      x-skaffolder-resource: Produto
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Produto
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Produto"
  "/produtos/findByIndustria/{key}":
    get:
      x-skaffolder-id: 61be9d597a739847d5cfd399
      x-skaffolder-name: findByIndustria
      x-skaffolder-id-resource: 61be9d077a739847d5cfd390
      x-skaffolder-resource: Produto
      x-skaffolder-crudAction: findByIndustria
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByIndustria
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByIndustria/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByIndustria
      tags:
        - Produto
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produto"
...
  /user:
    post:
      x-skaffolder-id: 61be5abf7a739847d5cfd2e8
      x-skaffolder-name: create
      x-skaffolder-id-resource: 61be5abf7a739847d5cfd2e0
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 61be5abf7a739847d5cfd2e7
      x-skaffolder-name: list
      x-skaffolder-id-resource: 61be5abf7a739847d5cfd2e0
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 61be5abf7a739847d5cfd2ea
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61be5abf7a739847d5cfd2e0
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61be5abf7a739847d5cfd2e3
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61be5abf7a739847d5cfd2e0
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 61be5abf7a739847d5cfd2e5
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61be5abf7a739847d5cfd2e0
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Produto:
      x-skaffolder-id: 61be9d077a739847d5cfd390
      x-skaffolder-id-db: 61be5abf7a739847d5cfd2d7
      x-skaffolder-id-entity: 61be9d077a739847d5cfd38d
      x-skaffolder-url: /produtos
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        nome:
          type: string
          x-skaffolder-id-attr: 61be9d077a739847d5cfd38e
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        Industria:
          x-skaffolder-id: 61be9d077a739847d5cfd38f
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 61be9d077a739847d5cfd38d
          x-skaffolder-ent2: 61be5d277a739847d5cfd36e
      required:
        - _id
        - nome
    User:
      x-skaffolder-id: 61be5abf7a739847d5cfd2e0
      x-skaffolder-id-db: 61be5abf7a739847d5cfd2d7
      x-skaffolder-id-entity: 61be5abf7a739847d5cfd2d8
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 61be5abf7a739847d5cfd2db
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 61be5abf7a739847d5cfd2dc
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 61be5abf7a739847d5cfd2da
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 61be5abf7a739847d5cfd2de
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 61be5abf7a739847d5cfd2dd
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 61be5abf7a739847d5cfd2d9
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...